Set ssh to false
I was getting `Rsync error rsync exited with code 255` with no clue what's going wrong, even if I've used the exact example from documtation. Downloaded and ran locally the node script and found out that it was requiring the password prompt because of `ssh: true`. Changed it to `false` and seems to work now.